\name{hiv.data}

\alias{hiv.data}

\docType{data}

\title{Example of an RtreemixData object}
\description{
    This data object was created by using the Stanford HIV Drug Resistance Database that comprises genetic
    measurements of 364 HIV patients treated only with the drug zidovudine.
    The data contains the six classical major zidovudine resistance mutations: M41L, D67N, K70R, L210W,
    T215F/Y, and K219E/Q. 
}

\usage{data(hiv.data)}

\seealso{
  \code{\link{RtreemixData-class}}
  }

\references{Human immunodeficiency virus reverse transcriptase and protease sequence
database, S. Rhee et al.}  

\examples{
data(hiv.data)

## print the object
hiv.data
}

\keyword{datasets}
